+++ src/writer/writer.service.ts
@@ -8,7 +8,6 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { DeleteResult, Repository } from 'typeorm';
 import { CreateWriterDto } from './dto/create-writer.dto';
 import { UpdateWriterDto } from './dto/update-writer.dto';
-import { validateObject } from 'src/validator/validate';
 import { ObjectId } from 'mongodb';
 
 @Injectable()
@@ -26,16 +25,12 @@ export class WriterService {
     const { email } = dto;
 
     const writer = await this.WriterRepository.findOneBy({ email: email });
-
     if (writer)
       throw new BadRequestException(`Writer with this email alerdy exist`);
 
-    let newWriter = new WriterEntity();
-    newWriter = Object.assign(newWriter, dto);
+    const newWriter = Object.assign(new WriterEntity(), dto);
     newWriter.role = 1;
     newWriter.books = [];
-    await validateObject(newWriter);
-
     return await this.WriterRepository.save(newWriter);
   }
 
@@ -46,16 +41,17 @@ export class WriterService {
   }
 
   async update(id: ObjectId, dto: UpdateWriterDto): Promise<WriterEntity> {
-    const writer = await this.WriterRepository.findOneBy({ _id: id });
+    const writer = await this.WriterRepository.findOneBy({ _id: id, role: 1 });
     if (!writer) throw new NotFoundException(`Writer with id ${id} not found`);
     delete writer.password;
-    delete writer.books;
 
     const updated = Object.assign(writer, dto);
     return await this.WriterRepository.save(updated);
   }
 
   async delete(id: ObjectId): Promise<DeleteResult> {
+    const writer = await this.WriterRepository.findBy({ _id: id, role: 1 });
+    if (!writer) throw new NotFoundException(`Writer with id ${id} not found`);
     return await this.WriterRepository.delete(id);
   }
 }
